+++ kernel/src/ipc/server/userspace.rs
@@ -25,7 +25,7 @@ impl UserspaceServer {
 }
 
 impl Server for UserspaceServer {
-	fn open(&self, endpoint: Cow<'_, Box<str>, str>) -> Result<u16, Error> {
+	fn open(&self, endpoint: Cow<'_, Box<str>, str>) -> Result<usize, Error> {
 		let endpoint = Box::<[u8]>::from(endpoint.into_owned()); // TODO(syscall-api): NUL terminate the string for better C interop
 		let packet = Packet {
 			proto_method: 0,
